게임 결과를 포맷해서 출력하는 기능을 MonsterGame클래스로 옮김
게임 결과 상태를 가지고 있는 MonsterGame에서 게임 결과를 출력하는 것이 좋다고 생각하여 기능을 옮김
메서드의 기능을 분리하고 테스트하기 쉽도록 게임 결과를 출력하는 메서드와 출력할 스트링을 반환하는 메서드를 분리함
